moved functions from deeptype.go to gentest.go
